create template for wizard
