+++ client/html/components/wizard.html
@@ -1,0 +1,3 @@
+<template name="wizard">
+  {{> Template.dynamic template=content}}
+</template>
