+++ client/html/pages/signup/students.html
@@ -1,4 +1,8 @@
 <template name="studentSignup">
+  {{> wizard }}
+</template>
+
+<template name="studentSteps">
   <div class="card signup-card">
     <div class="card-body">
       <div id="wizard" class="form-wizard form-wizard-horizontal">
